+++ include/bencode.h
@@ -302,7 +302,7 @@ INLINE bencode_item_t *bencode_decode_expect(bencode_buffer_t *buf, const char *
 INLINE bencode_item_t *bencode_decode_expect_str(bencode_buffer_t *buf, const str *s, bencode_type_t expect);
 
 /* Returns the number of bytes that could successfully be decoded from 's', -1 if more bytes are needed or -2 on error */
-int bencode_valid(const char *s, size_t len);
+ssize_t bencode_valid(const char *s, size_t len);
 
 
 /*** DICTIONARY LOOKUP & EXTRACTION ***/
